Move the func folders; update requirements
